# restapi/DMCInterfaceSwagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: DMC Interface Definition
  description: |
    **REST service interface definitions for DMC**
host: dmc.com
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
  - text/xml
produces:
  - application/json
  - text/html
paths:
  /user:
    get:
      tags:
        - account
        - live
        - has tests
      description: Get currently logged in user details.  If user does not exist, create account.
      responses:
        '200':
          description: Sends the project with projectID
          schema:
            $ref: '#/definitions/UserDetails'
    patch:
      tags:
      - account
      - onboarding
      - live
      - has tests
      description: Update currently logged in user details
      parameters:
        - name: UserDetails
          in: body
          schema:
            $ref: '#/definitions/UserDetails'
      responses:
        200:
          description: returns updated User Details
          schema:
            $ref: '#/definitions/UserDetails'
...
  '/mark-read-notifications':
    patch:
      tags:
        - notification
      description: Mark read for all user notifications
      parameters:
        - name: UserDetails
          in: body
          schema:
            $ref: '#/definitions/UserDetails'
          required: true
      responses:
        200:
          description: returns updated User Details
          schema:
            $ref: '#/definitions/UserDetails'
  '/clear-notification/{id}':
    patch:
      tags:
        - notification
      description: Clear notification from dropdown in header
      parameters:
        - name: id
          in: path
          type: string
          description: ID notification
          required: true
        - name: UserDetails
          in: body
          schema:
            $ref: '#/definitions/UserDetails'
          required: true
      responses:
        200:
          description: returns updated User Details
          schema:
            $ref: '#/definitions/UserDetails'
...
  '/project/{projectId}/invite/{projectMemberId}':
    patch:
      tags:
        - project
      description: Re Invite user to the project from team project page (from removed tab)
      parameters:
        - name: projectId
          in: path
          type: string
          description: ID of the project
          required: true
        - name: projectMemberId
          in: path
          type: string
          description: ID of the project member
          required: true
        - in: body
          name: Member
          description: Member to be updated
          required: false
          schema:
            $ref: '#/definitions/ProjectMember'
      responses:
        '200':
          description: return project member
          schema:
            $ref: '#/definitions/ProjectMember'
...
  '/projects/all':
    get:
      tags:
        - project
      description: Get all projects
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
        - name: title_like
          in: query
          description: Filter by title
          type: string
          required: false  
      responses:
        '200':
          description: return array of projects
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
  '/projects':
    get:
      tags:
        - project
        - live
        - has tests
      description: Get all user projects
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
        - name: title_like
          in: query
          description: Filter by title
          type: string
      responses:
        '200':
          description: return array of projects
          schema:
            type: array
            items:
              $ref: '#/definitions/Project'
  /projects/create:
    post:
      tags:
        - project
        - live
        - has tests
      description: Post a new project to store in the database
      operationId: postProject
      parameters:
        - in: body
          name: body
          description: Project object to be added to the database
          required: false
          schema:
            $ref: '#/definitions/Project'
      responses:
        '201':
          description: New project
          schema:
            $ref: '#/definitions/Project'
...
  '/projects/{projectID}/following_discussions':
    get:
      tags:
        - project
        - discussion
        - not implemented
      description: Get all project discussions being followed
      parameters:
        - name: projectID
          in: path
          type: string
          description: ID of the project
          required: true
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
      responses:
        '200':
          description: return all project discussions being followed
          schema:
            type: array
            items:
              $ref: '#/definitions/IndividualDiscussion'
...
  '/projects_members/{memberId}':
    get:
      tags:
        - project
        - add member
        - not implemented
      description: Get member by id
      parameters:
        - in: path
          name: memberId
          description: member id
          required: true
          type: string
      responses:
        '200':
          description: return member
          schema:
            $ref: '#/definitions/ProjectMember'
  '/projects/{projectId}/accept/{memberId}':
    patch:
      tags:
        - project
        - add member
        - not implemented
      description: Accept member inivitation
      parameters:
        - in: path
          name: memberId
          type: string
          description: Project member id
          required: true
        - in: path
          name: projectId
          type: string
          description: Project id
          required: true
        - in: body
          name: Member
          schema:
            $ref: '#/definitions/ProjectMember'
      responses:
        '200':
          description: Return updated member
          schema:
            $ref: '#/definitions/ProjectMember'
  '/projects/{projectId}/reject/{memberId}':
    delete:
      tags:
        - project
        - add member
        - not implemented
      description: Decline member inivitation
      parameters:
        - in: path
          name: memberId
          type: string
          description: member id
          required: true
        - in: path
          name: projectId
          type: string
          description: Project id
          required: true
      responses:
        '200':
          description: Ok
  /tasks/create:
    post:
      tags:
        - task
        - live
        - has tests
      summary: Post a new task
      description: Post a new task to store in the database
      operationId: postTask
      parameters:
        - in: body
          name: body
          description: Task object to be added to the database
          required: false
          schema:
            $ref: '#/definitions/Task'
      responses:
        '200':
          description: return created task
          schema:
            $ref: '#/definitions/Task'
        '405':
          description: Invalide input
  '/tasks':
    get:
      tags:
        - task
        - live
        - has tests
      summary: Get tasks
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
        - name: title_like
          in: query
          description: Select by title name
          type: string
          required: false
      responses:
        '200':
          description: send the task with taskID
          schema:
            type: array
            items:
              $ref: '#/definitions/Task'
...
  '/services':
    post:
      tags:
        - service
        - not implemented
      summary: Post a new service
      description: Post a new service to store in the database
      operationId: postService
      parameters:
        - in: body
          name: body
          description: Service object to be added to the database
          required: false
          schema:
            $ref: '#/definitions/Service'
      responses:
        '405':
          description: Invalide input
    get:
      tags:
        - service
        - marketplace
        - live
        - has tests
      description: get all services
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
      responses:
        '200':
          description: send all services
          schema:
            type: array
            items:
              $ref: '#/definitions/Service'
...
  /components:
    post:
      tags:
        - component
        - not implemented
      summary: Post a new component
      description: Post a new component to store in the database
      operationId: postComponent
      parameters:
        - in: body
          name: body
          description: Component object to be added to the database
          required: false
          schema:
            $ref: '#/definitions/Component'
      responses:
        '405':
          description: Invalide input
...
  '/popular_discussions':
    get:
      tags:
        - discussion
        - not implemented
      description: Get popular discussions
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
      responses:
        '200':
          description: return popular discussions
          schema:
            type: array
            items:
              $ref: '#/definitions/IndividualDiscussion'
  '/follow_people_discussions':
    get:
      tags:
        - discussion
        - not implemented
      description: Get discussion of people being followed
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
      responses:
        '200':
          description: return discussions of people being
          schema:
            type: array
            items:
              $ref: '#/definitions/IndividualDiscussion'
...
  /discussions/create:
    post:
      tags:
        - discussion
        - live
        - has tests
      summary: Post a new discussion
      description: Post a new discussion to store in the database
      operationId: postDiscussion
      parameters:
        - in: body
          name: body
          description: Discussion object to be added to the database
          required: false
          schema:
            $ref: '#/definitions/IndividualDiscussion'
      responses:
        '405':
          description: Invalide input
  '/profiles':
    post:
      tags:
        - profile
        - live
        - has tests
      description: Create a new user profile
      parameters:
        - in: body
          name: Profile
          description: Profile to be created
          required: false
          schema:
            $ref: '#/definitions/Profile'
      responses:
        '201':
          description: New user profile created
          schema:
            $ref: '#/definitions/Profile'
    get:
      tags:
        - profile
        - not implemented
      description: Get all user profiles
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter
          type: string
          required: false
        - name: id
          in: query
          description: Select by profile id
          type: array
          items:
            type: string
          required: false
      responses:
        '200':
          description: send all user profiles
          schema:
            type: array
            items:
              $ref: '#/definitions/Profile'
...
  '/companies/{companyID}/company_services':
    get:
      tags:
        - company
        - service
        - not implemented
      description: Get company services based on companyID
      parameters:
        - name: companyID
          in: path
          type: string
          description: ID of the company
          required: true
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _start
          in: query
          description: Offset
          type: integer
          required: false
        - name: _sort
          in: query
          description: Sort by some parameter (id, title, etc.)
          type: string
          required: false
        - name: title_like
          in: query
          description: Search by service title
          type: string
          required: false
        - name: serviceType
          in: query
          description: Filter by service type (for service)
          type: string
          required: false
          collectionFormat: multi
          enum:
            - analytical
            - solid
            - data
        - name: _authors
          in: query
          description: Filter by author
          type: array
          items:
            type: integer
          required: false
        - name: _ratings
          in: query
          description: Filter by rating
          type: array
          items:
            type: string
          required: false
        - name: _favorites
          in: query
          description: Filter by favorite items
          type: string
          required: false
        - name: _dates
          in: query
          description: Filter by date (7d - 7 days old, 1m - 1 month old, 1y - 1 year old)
          type: array
          items:
            type: string
          required: false
          collectionFormat: multi
          enum:
            - 7d
            - 1m
            - 1y
      responses:
        '200':
          description: send the services with companyID
          schema:
            type: array
            items:
              $ref: '#/definitions/Service'
  '/companies/{companyID}/company_components':
    get:
      tags:
        - company
        - component
        - not implemented
      description: Get company components based on companyID
      parameters:
        - name: companyID
          in: path
          type: string
          description: ID of the company
          required: true
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _start
          in: query
          description: Offset
          type: integer
          required: false
        - name: _sort
          in: query
          description: Sort by some parameter (id, title, etc.)
          type: string
          required: false
        - name: title_like
          in: query
          description: Search by service title
          type: string
          required: false
        - name: serviceType
          in: query
          description: Filter by service type (for service)
          type: string
          required: false
          collectionFormat: multi
          enum:
            - analytical
            - solid
            - data
        - name: _authors
          in: query
          description: Filter by author
          type: array
          items:
            type: integer
          required: false
        - name: _ratings
          in: query
          description: Filter by rating
          type: array
          items:
            type: string
          required: false
        - name: _favorites
          in: query
          description: Filter by favorite items
          type: string
          required: false
        - name: _dates
          in: query
          description: Filter by date (7d - 7 days old, 1m - 1 month old, 1y - 1 year old)
          type: array
          items:
            type: string
          required: false
          collectionFormat: multi
          enum:
            - 7d
            - 1m
            - 1y
      responses:
        '200':
          description: send the components with companyID
          schema:
            type: array
            items:
              $ref: '#/definitions/Component'
...
  '/companies/{companyID}/new':
    get:
      tags:
        - company
        - not implemented
      description: Get new company services
      parameters:
        - name: companyID
          in: path
          type: string
          description: ID of the company
          required: true
        - name: _order
          in: query
          description: Order by DESC
          type: string
          collectionFormat: multi
          required: true
          enum:
            - DESC
        - name: _sort
          in: query
          description: Sort by id
          type: string
          required: true
        - name: _limit
          in: query
          description: Limit services
          type: string
          required: true
      responses:
        '200':
          description: send an object with new services based on companyID
          schema:
            type: array
            items:
              $ref: '#/definitions/Service'
...
  '/company/follow':
    post:
      tags:
        - company
        - not implemented
      description: Follow company
      parameters:
        - name: accountId
          type: integer
          description: User account id
          in: query
          required: true
        - name: companyId
          type: integer
          description: Company id
          in: query
          required: true
      responses:
        '201':
          description: New followed company
          schema:
            type: object
            properties:
              id:
                type: integer
              accountId:
                type: integer
              companyId:
                type: integer
  '/company/unfollow/{followed_companiId}':
    delete:
      tags:
        - company
        - not implemented
      description: Unfollow company
      parameters:
        - in: path
          name: followed_companiId
          type: integer
          description: Followed company id
          required: true
      responses:
        '200':
          description: OK
  '/company_featured/{company_featuredId}/position':
    patch:
      tags:
        - company
        - not implemented
      description: Update featured position
      parameters:
        - in: path
          name: company_featuredId
          type: integer
          description: Company id
          required: true
        - in: query
          name: position
          type: integer
          description: new company featured position
          required: true
      responses:
        '202':
          description: Updated featured company
          schema:
            type: object
            properties:
              id:
                type: integer
              companyId:
                type: integer
              position:
                type: integer
              serviceId:
                type: integer
  '/company_featured/add':
    post:
      tags:
        - company
        - not implemented
      description: Added new featured company
      parameters:
        - in: query
          name: companyId
          type: integer
          description: Company id
          required: true
        - in: query
          name: position
          type: integer
          description: Featured position
          required: true
        - in: query
          name: serviceId
          type: integer
          description: Service id
          required: true
      responses:
        '201':
          description: New featured company
          schema:
            type: object
            properties:
              id:
                type: integer
              companyId:
                type: integer
              position:
                type: integer
              serviceId:
                type: integer
  '/company_featured/{company_featuredId}':
    delete:
      tags:
        - company
        - not implemented
      description: Delete featured company
      parameters:
        - in: path
          name: company_featuredId
          type: integer
          description: Featured company id
          required: true
      responses:
        '200':
          description: OK
  '/market/popular_services':
    get:
      tags:
        - service
        - marketplace
        - not implemented
      description: Get popular services in marketplace
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _start
          in: query
          description: Offset
          type: integer
          required: false
        - name: _sort
          in: query
          description: Sort by some parameter (id, title, etc.)
          type: string
          required: false
      responses:
        '200':
          description: return popular services in marketplace
          schema:
            type: array
            items:
              $ref: '#/definitions/Service'
  '/market/new_services':
    get:
      tags:
        - service
        - marketplace
        - not implemented
      description: Get latest services in marketplace
      parameters:
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _start
          in: query
          description: Offset
          type: integer
          required: false
        - name: _sort
          in: query
          description: Sort by some parameter (id, title, etc.)
          type: string
          required: false
      responses:
        '200':
          description: return latest services in marketplace
          schema:
            type: array
            items:
              $ref: '#/definitions/Service'
...
  '/follow-company-services':
    get:
      tags:
        - service
        - marketplace
        - not implemented
      description: Get all services from favorite storefronts
      parameters:
        - name: companyId
          in: query
          description: Array of favorite companies id
          type: array
          items:
            type: integer
          required: true
        - name: _limit
          in: query
          description: Limit items
          type: integer
          required: false
        - name: _start
          in: query
          description: Offset
          type: integer
          required: false
        - name: _order
          in: query
          description: Order by ASC or DESC
          type: string
          collectionFormat: multi
          required: false
          enum:
            - ASC
            - DESC
        - name: _sort
          in: query
          description: Sort by some parameter (id, title, etc.)
          type: string
          required: false
      responses:
        '200':
          description: return all services from favorite storefronts
          schema:
            type: array
            items:
              $ref: '#/definitions/Service'
...
definitions:
...
  FeatureImage:
    type: object
    properties:
      thumbnail:
        type: string
      large:
        type: string
...
  Project:
    type: object
    properties:
      id:
        type: string
      title:
        type: string
      projectManager:
        type: string
      projectManagerId:
        type: string
      companyId:
        type: string
      featureImage:
        $ref: '#/definitions/FeatureImage'
      images:
        type: string
      description:
        type: string
      dueDate:
        type: number
        format: milliseconds
      tasks:
        $ref: '#/definitions/Tasks'
      discussions:
        $ref: '#/definitions/Discussions'
      services:
        $ref: '#/definitions/Services'
      components:
        $ref: '#/definitions/Components'
      approvalOption:
        type: string
...
  UserDetails:
    type: object
    properties:
      displayName:
        type: string
      accountId:
        type: string
      profileId:
        type: string
      companyId:
        type: string
      role:
        type: string
      termsConditions:
        type: boolean
      notifications:
        type: object
        properties:
          total:
            type: number
          items:
            type: array
            items:
              $ref: '#/definitions/UserNotification'
      runningServices:
        type: object
        properties:
          total:
            type: number
          items:
            type: array
            items:
              $ref: '#/definitions/RunningServicesListItem'

      messages:
        type: object
        properties:
          total:
            type: integer
          items:
            type: array
            items:
              $ref: '#/definitions/MessagesListItem'
      onboarding:
        type: object
        properties:
          profile:
            type: boolean
          account:
            type: boolean
          company:
            type: boolean
          storefront:
            type: boolean
...